features: - generate the irradiance map - add bilinear interpolation - add operator-() to Vec3 and Vec4 refactor: - refactor texture and bmp_loader

+++ SoftRasterLearning/render_test/render_test_normal_mapping.hpp
@@ -149,7 +149,7 @@ public:
 	virtual void Update(const framework::IRenderEngine& engine) override
 	{
 		size_t delta = engine.GetEngineState().delta_count;
-		sphere->transform.rotation += core::Vec3{ 0.f, 1.f, 0.f }*0.001f * delta;
+		sphere->transform.rotation += core::Vec3{ 0.f, 1.f, 0.f }*0.001f * (float)delta;
 	}
 
 	virtual const framework::ICamera* GetMainCamera() const override
